+++ client/src/components/Chart.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { useSelector } from "react-redux";
+
+
+const NewMap = () => {
+  
+
+  return (
+    <div className="newMap">
+    </div>
+  );
+};
+
+export default NewMap;
