+++ client/src/components/Cost.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { useSelector } from "react-redux";
+
+
+const Cost = () => {
+  
+
+  return (
+    <div className="cost">
+    </div>
+  );
+};
+
+export default Cost;
